formatting sol files

+++ script/Counter.s.sol
@@ -1,16 +1,16 @@
 // SPDX-License-Identifier: UNLICENSED
 pragma solidity ^0.8.13;
 
-import {Script, console} from "forge-std/Script.sol";
-import {SimpleCounter} from "../src/SimpleCounter.sol";
+import { Script, console } from "forge-std/Script.sol";
+import { SimpleCounter } from "../src/SimpleCounter.sol";
 
 contract SimpleCounterScript is Script {
     SimpleCounter public counter;
 
-    function setUp() public {}
+    function setUp() public { }
 
     function run() public {
-        uint privateKey = vm.envUint("WALLET_PRIVATE_KEY");
+        uint256 privateKey = vm.envUint("WALLET_PRIVATE_KEY");
         address addr = vm.addr(privateKey);
         console.log("address", addr);
         console.log("Chain ID:", block.chainid);
